@@ -35,17 +35,17 @@
     "prepare": "run-s prepare:*"
   },
   "devDependencies": {
-    "@arcticicestudio/eslint-config-base": ">=0.11.0 <=1.0.0",
-    "@arcticicestudio/remark-preset-lint": ">=0.4.0 <1.0.0",
+    "@svengreb/eslint-config-base": ">=0.12.0 <=1.0.0",
+    "@svengreb/remark-preset-lint": ">=0.5.0 <1.0.0",
     "@prettier/plugin-xml": "2.2.0",
-    "eslint": "8.38.0",
+    "eslint": "8.39.0",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-prettier": "4.2.1",
     "husky": "8.0.3",
-    "lint-staged": "13.2.1",
+    "lint-staged": "13.2.2",
     "npm-run-all": "4.1.5",
-    "prettier": "2.8.7",
+    "prettier": "2.8.8",
     "prettier-plugin-sh": "0.12.8",
     "remark-cli": "11.0.0"
   }
