@@ -8,6 +8,6 @@
  * @see https://github.com/remarkjs/remark-lint
  * @see https://remark.js.org
  */
-module.exports = {
-  plugins: ["@arcticicestudio/remark-preset-lint"],
+export default {
+  plugins: ["@svengreb/remark-preset-lint"],
 };
