@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2016-present Sven Greb <development@svengreb.de>
- * This source code is licensed under the MIT license found in the license file.
+ * This source code is licensed under the Creative Commons Attribution-ShareAlike 4.0 International license found in the license file.
  */
 
 /**
@@ -12,7 +12,15 @@
  */
 module.exports = {
   root: true,
-  extends: ["@arcticicestudio/eslint-config-base", "@arcticicestudio/eslint-config-base/prettier"],
+  extends: [
+    "@svengreb/eslint-config-base",
+    /*
+     * Enable support for projects using Prettier.
+     * Note that this must always be placed after the `@svengreb/eslint-config-base` preset to take precedence, otherwise it won't prevent errors
+     * due to useless and possibly conflicting rules!
+     */
+    "@svengreb/eslint-config-base/prettier",
+  ],
   overrides: [
     {
       files: ["*.js"],
